# tests/contrato_openapi_v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UsuarioGestionActualizarPeticion": {
        "properties": {
          "apellidos": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Apellidos",
            "type": "string"
          },
          "correo": {
            "format": "email",
            "title": "Correo",
            "type": "string"
          },
          "estado": {
            "default": 1,
            "title": "Estado",
            "type": "integer"
          },
          "foto_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foto Url"
          },
          "nombre": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Nombre",
            "type": "string"
          },
          "rol_id": {
            "title": "Rol Id",
            "type": "integer"
          },
          "telefono": {
            "anyOf": [
              {
                "maxLength": 12,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          }
        },
        "required": [
          "nombre",
          "apellidos",
          "correo",
          "rol_id"
        ],
        "title": "UsuarioGestionActualizarPeticion",
        "type": "object"
      },
      "UsuarioGestionCrearPeticion": {
        "properties": {
          "apellidos": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Apellidos",
            "type": "string"
          },
          "correo": {
            "format": "email",
            "title": "Correo",
            "type": "string"
          },
          "estado": {
            "default": 1,
            "title": "Estado",
            "type": "integer"
          },
          "foto_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foto Url"
          },
          "nombre": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Nombre",
            "type": "string"
          },
          "rol_id": {
            "title": "Rol Id",
            "type": "integer"
          },
          "telefono": {
            "anyOf": [
              {
                "maxLength": 12,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          }
        },
        "required": [
          "nombre",
          "apellidos",
          "correo",
          "rol_id"
        ],
        "title": "UsuarioGestionCrearPeticion",
        "type": "object"
      },
      "UsuarioGestionEstadoPeticion": {
        "properties": {
          "estado": {
            "title": "Estado",
            "type": "integer"
          }
        },
        "required": [
          "estado"
        ],
        "title": "UsuarioGestionEstadoPeticion",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/autenticacion/iniciar-sesion"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Arquetipo Idiomatico Escalable de FastAPI",
    "title": "FastAPI Arquetipo",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/usuario/perfil": {
      "get": {
        "operationId": "obtener_mi_perfil_api_usuario_perfil_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Obtener Mi Perfil",
        "tags": [
          "Usuario"
        ]
      }
    },
    "/api/usuario/{usuario_id}/perfil": {
      "get": {
        "operationId": "obtener_perfil_por_usuario_id_api_usuario__usuario_id__perfil_get",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "title": "Usuario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Obtener Perfil Por Usuario Id",
        "tags": [
          "Usuario"
        ]
      }
    },
    "/api/usuarios": {
      "get": {
        "operationId": "listar_usuarios_gestion_api_usuarios_get",
        "parameters": [
          {
            "in": "query",
            "name": "pagina",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Pagina",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tamano",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Tamano",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "correo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Correo"
            }
          },
          {
            "in": "query",
            "name": "rol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rol"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Listar Usuarios Gestion",
        "tags": [
          "Usuarios Gestion"
        ]
      },
      "post": {
        "operationId": "crear_usuario_gestion_api_usuarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioGestionCrearPeticion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Crear Usuario Gestion",
        "tags": [
          "Usuarios Gestion"
        ]
      }
    },
    "/api/usuarios/{usuario_id}": {
      "delete": {
        "operationId": "eliminar_logico_usuario_gestion_api_usuarios__usuario_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "title": "Usuario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Eliminar Logico Usuario Gestion",
        "tags": [
          "Usuarios Gestion"
        ]
      },
      "get": {
        "operationId": "obtener_usuario_gestion_api_usuarios__usuario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "title": "Usuario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Obtener Usuario Gestion",
        "tags": [
          "Usuarios Gestion"
        ]
      },
      "put": {
        "operationId": "actualizar_usuario_gestion_api_usuarios__usuario_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "title": "Usuario Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioGestionActualizarPeticion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Actualizar Usuario Gestion",
        "tags": [
          "Usuarios Gestion"
        ]
      }
    },
    "/api/usuarios/{usuario_id}/estado": {
      "patch": {
        "operationId": "cambiar_estado_usuario_gestion_api_usuarios__usuario_id__estado_patch",
        "parameters": [
          {
            "in": "path",
            "name": "usuario_id",
            "required": true,
            "schema": {
              "title": "Usuario Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioGestionEstadoPeticion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Cambiar Estado Usuario Gestion",
        "tags": [
          "Usuarios Gestion"
        ]
      }
    }
  }
}